+++ src/main/java/org/advent/helper/Point.java
@@ -8,6 +8,6 @@ import lombok.Data;
 @Builder
 @AllArgsConstructor
 public class Point {
-    private int x;
-    private int y;
+    private double x;
+    private double y;
 }
